finalize component conversion
